Microsoft word support bot.
The project idea is to solve the problem which new user face in Microsoft word.
I'm trying to solve the basic problem which arises due to lack of the skills to search in web or it is time consuming, as MS word functions are available in the official Microsoft guide to word as well but it is too much information in one go and it is pretty over whelming, so to clearly provide the information the bot will fill this gap with answers as short as one liner solutions.
As clearly stated above there is a huge gap between the application an the user knowledge about it, so to fill this gap a full conversational bot will come into play first it will be the based on simple commands like just write your need in one or tow words like function or any other service the bot will reply to most appropriate answer in later time it will even correct your mistakes in the command or words you supply.